renames add_keys to scan_range, moves executor to scan_task mod
