+++ zebra-scan/src/service/scan_task/scan/scan_range.rs
